#97: Complete refactoring with best practices  (#105)
* Remove context from shareImage
* Delete getCachedImageUri
* Extract image loading from EditViewModel file
* Move all state to EditingState
* Clarify Serialization in ImageDefaults
* Fix Share
