+++ app/src/main/java/dev/arkbuilders/arkretouch/App.kt
@@ -2,8 +2,8 @@ package dev.arkbuilders.arkretouch
 
 import android.app.Application
 import dev.arkbuilders.arkfilepicker.folders.FoldersRepo
-import dev.arkbuilders.arkretouch.common.CommonModule
-import dev.arkbuilders.arkretouch.edition.EditModule
+import dev.arkbuilders.arkretouch.di.EditModule
+import dev.arkbuilders.arkretouch.di.common.CommonModule
 import org.acra.config.dialog
 import org.acra.config.httpSender
 import org.acra.data.StringFormat
