--- app/src/main/java/dev/arkbuilders/arkretouch/common/InjectionModule.kt
@@ -1,7 +1,0 @@
-package com.etherean.app.common.di
-
-import org.koin.core.module.Module
-
-fun interface InjectionModule {
-    fun create(): Module
-}
