+++ app/src/main/java/dev/arkbuilders/arkretouch/di/common/InjectionModule.kt
@@ -1,0 +1,7 @@
+package dev.arkbuilders.arkretouch.di.common
+
+import org.koin.core.module.Module
+
+fun interface InjectionModule {
+    fun create(): Module
+}
