+++ app/src/main/java/dev/arkbuilders/arkretouch/di/common/dispatchers/CoroutinesDispatchers.kt
@@ -1,0 +1,19 @@
+package dev.arkbuilders.arkretouch.di.common.dispatchers
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+
+interface CoroutineDispatchers {
+    val io: CoroutineDispatcher
+    val computation: CoroutineDispatcher
+    val ui: CoroutineDispatcher
+}
+
+class DefaultDispatchers : CoroutineDispatchers {
+    override val io: CoroutineDispatcher
+        get() = Dispatchers.IO
+    override val computation: CoroutineDispatcher
+        get() = Dispatchers.Default
+    override val ui: CoroutineDispatcher
+        get() = Dispatchers.Main.immediate
+}
