+++ app/src/main/java/dev/arkbuilders/arkretouch/editing/manager/EditingMode.kt
@@ -1,0 +1,9 @@
+package dev.arkbuilders.arkretouch.editing.manager
+
+enum class EditingMode {
+    IDLE,
+    CROP,
+    RESIZE,
+    ROTATE,
+    ZOOM
+}
