+++ app/src/main/java/dev/arkbuilders/arkretouch/editing/model/DrawPath.kt
@@ -1,0 +1,9 @@
+package dev.arkbuilders.arkretouch.editing.model
+
+import androidx.compose.ui.graphics.Paint
+import androidx.compose.ui.graphics.Path
+
+class DrawPath(
+    val path: Path,
+    val paint: Paint
+)
