+++ app/src/main/java/dev/arkbuilders/arkretouch/editing/model/ImageViewParams.kt
@@ -1,0 +1,9 @@
+package dev.arkbuilders.arkretouch.editing.model
+
+import androidx.compose.ui.unit.IntSize
+import dev.arkbuilders.arkretouch.presentation.edit.resize.ResizeOperation
+
+class ImageViewParams(
+    val drawArea: IntSize,
+    val scale: ResizeOperation.Scale
+)
