+++ app/src/main/java/dev/arkbuilders/arkretouch/editing/model/Operation.kt
@@ -1,0 +1,9 @@
+package dev.arkbuilders.arkretouch.editing.model
+
+interface Operation {
+    fun apply()
+
+    fun undo()
+
+    fun redo()
+}
