+++ app/src/main/java/dev/arkbuilders/arkretouch/editing/ui/canvas/EditCanvas.kt
@@ -1,0 +1,110 @@
+package dev.arkbuilders.arkretouch.editing.ui.canvas
+
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.gestures.awaitFirstDown
+import androidx.compose.foundation.gestures.calculatePan
+import androidx.compose.foundation.gestures.calculateZoom
+import androidx.compose.foundation.gestures.forEachGesture
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.input.pointer.pointerInput
+import dev.arkbuilders.arkretouch.editing.ui.main.EditViewModel
+import dev.arkbuilders.arkretouch.picker.toDp
+import dev.arkbuilders.arkretouch.presentation.edit.TransparencyChessBoardCanvas
+import dev.arkbuilders.arkretouch.utils.calculateRotationFromOneFingerGesture
+
+@Composable
+fun EditCanvasScreen(viewModel: EditViewModel) {
+    val editManager = viewModel.editManager
+    var scale by remember { mutableStateOf(1f) }
+    var offset by remember { mutableStateOf(Offset.Zero) }
+
+    fun resetScaleAndTranslate() {
+        editManager.apply {
+            if (
+                isRotateMode.value || isCropMode.value || isResizeMode.value ||
+                isBlurMode.value
+            ) {
+                scale = 1f; zoomScale = scale; offset = Offset.Zero
+            }
+        }
+    }
+
+    Box(contentAlignment = Alignment.Center) {
+        val modifier = Modifier
+            .size(
+                editManager.availableDrawAreaSize.value.width.toDp(),
+                editManager.availableDrawAreaSize.value.height.toDp()
+            )
+            .graphicsLayer {
+                resetScaleAndTranslate()
+
+                // Eraser leaves black line instead of erasing without this hack, it uses BlendMode.SrcOut
+                // https://stackoverflow.com/questions/65653560/jetpack-compose-applying-porterduffmode-to-image
+                // Provide a slight opacity to for compositing into an
+                // offscreen buffer to ensure blend modes are applied to empty pixel information
+                // By default any alpha != 1.0f will use a compositing layer by default
+                alpha = 0.99f
+
+                scaleX = scale
+                scaleY = scale
+                translationX = offset.x
+                translationY = offset.y
+            }
+        TransparencyChessBoardCanvas(modifier, editManager)
+        BackgroundCanvas(modifier, editManager)
+        DrawCanvas(modifier, viewModel)
+    }
+    if (
+        editManager.isRotateMode.value || editManager.isZoomMode.value ||
+        editManager.isPanMode.value
+    ) {
+        Canvas(
+            Modifier
+                .fillMaxSize()
+                .pointerInput(Any()) {
+                    forEachGesture {
+                        awaitPointerEventScope {
+                            awaitFirstDown()
+                            do {
+                                val event = awaitPointerEvent()
+                                when (true) {
+                                    (editManager.isRotateMode.value) -> {
+                                        val angle = event
+                                            .calculateRotationFromOneFingerGesture(
+                                                editManager.calcCenter()
+                                            )
+                                        editManager.rotate(angle)
+                                        editManager.invalidatorTick.value++
+                                    }
+                                    else -> {
+                                        if (editManager.isZoomMode.value) {
+                                            scale *= event.calculateZoom()
+                                            editManager.zoomScale = scale
+                                        }
+                                        if (editManager.isPanMode.value) {
+                                            val pan = event.calculatePan()
+                                            offset = Offset(
+                                                offset.x + pan.x,
+                                                offset.y + pan.y
+                                            )
+                                        }
+                                    }
+                                }
+                            } while (event.changes.any { it.pressed })
+                        }
+                    }
+                }
+        ) {}
+    }
+}
