--- app/src/main/java/dev/arkbuilders/arkretouch/edition/manager/EditionMode.kt
@@ -1,9 +1,0 @@
-package dev.arkbuilders.arkretouch.edition.manager
-
-enum class EditionMode {
-    IDLE,
-    CROP,
-    RESIZE,
-    ROTATE,
-    ZOOM
-}
