--- app/src/main/java/dev/arkbuilders/arkretouch/edition/model/DrawPath.kt
@@ -1,9 +1,0 @@
-package dev.arkbuilders.arkretouch.edition.model
-
-import androidx.compose.ui.graphics.Paint
-import androidx.compose.ui.graphics.Path
-
-class DrawPath(
-    val path: Path,
-    val paint: Paint
-)
