--- app/src/main/java/dev/arkbuilders/arkretouch/edition/model/EditionState.kt
@@ -1,12 +1,0 @@
-package dev.arkbuilders.arkretouch.edition.model
-
-data class EditionState(
-    val strokeSliderExpanded: Boolean
-) {
-
-    companion object {
-        val DEFAULT = EditionState(
-            strokeSliderExpanded = false
-        )
-    }
-}
