--- app/src/main/java/dev/arkbuilders/arkretouch/edition/model/Operation.kt
@@ -1,9 +1,0 @@
-package dev.arkbuilders.arkretouch.edition.model
-
-interface Operation {
-    fun apply()
-
-    fun undo()
-
-    fun redo()
-}
