+++ app/src/main/java/dev/arkbuilders/arkretouch/presentation/edit/TransparencyChessBoard.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
 import androidx.compose.ui.graphics.nativeCanvas
 import androidx.compose.ui.unit.toSize
 import android.graphics.Matrix
-import dev.arkbuilders.arkretouch.edition.manager.EditManager
+import dev.arkbuilders.arkretouch.editing.manager.EditManager
 
 private class TransparencyChessBoard {
     fun create(boardSize: Size, canvas: Canvas, matrix: Matrix) {
