+++ app/src/main/java/dev/arkbuilders/arkretouch/presentation/edit/resize/ResizeOperation.kt
@@ -5,8 +5,8 @@ import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.unit.IntSize
 import android.graphics.Bitmap
-import dev.arkbuilders.arkretouch.edition.manager.EditManager
-import dev.arkbuilders.arkretouch.edition.model.Operation
+import dev.arkbuilders.arkretouch.editing.manager.EditManager
+import dev.arkbuilders.arkretouch.editing.model.Operation
 import dev.arkbuilders.arkretouch.utils.resize
 import java.lang.NullPointerException
