+++ app/src/main/java/dev/arkbuilders/arkretouch/presentation/edit/rotate/RotateOperation.kt
@@ -1,8 +1,8 @@
 package dev.arkbuilders.arkretouch.presentation.edit.rotate
 
 import android.graphics.Matrix
-import dev.arkbuilders.arkretouch.edition.manager.EditManager
-import dev.arkbuilders.arkretouch.edition.model.Operation
+import dev.arkbuilders.arkretouch.editing.manager.EditManager
+import dev.arkbuilders.arkretouch.editing.model.Operation
 import dev.arkbuilders.arkretouch.utils.rotate
 
 class RotateOperation(private val editManager: EditManager) : Operation {
