+++ app/src/main/java/dev/arkbuilders/arkretouch/storage/ImageDefaults.kt
@@ -4,7 +4,12 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.IntSize
 import kotlinx.serialization.Serializable
 
-// FIXME: Replace with Parcelable
+/**
+ * We use serialization for these classes because we write them to storage as JSON text,
+ * To convert them correctly we use [kotlinx.serialization.encodeToString] and [kotlinx.serialization.decodeFromString] which requires an object to be [@Serializable]
+ * Refer here [dev.arkbuilders.arkretouch.storage.OldStorageRepository.persistDefaults]
+*/
+
 @Serializable
 data class ImageDefaults(
     val colorValue: ULong = Color.White.value,
